Add basic support for connecting and querying data from MySQL database. Add d3node-linechart package.
This commit contains a prrof of concept for using D3 and data from MySQL. When the webpage
receives a get request, it reads the data not from the sample csv, but from a MySQL
connection set up.

Note: the code contains many shortcuts that allow it to work. It needs to be refactored.